+++ internal/usecase/address_usecase_test.go
@@ -16,7 +16,7 @@ func TestAddressUsecase_CreateAddress_Success(t *testing.T) {
 	mockRegionService := new(mocks.MockRegionService)
 	addressUsecase := NewAddressUsecase(mockAddressRepo, mockRegionService)
 
-	userID := uint(1)
+	userID := uint64(1)
 	req := &domain.CreateAddressRequest{
 		Name:       "Rumah",
 		Detail:     "Jl. Sudirman No. 123",
@@ -59,7 +59,7 @@ func TestAddressUsecase_CreateAddress_InvalidRegion(t *testing.T) {
 	mockRegionService := new(mocks.MockRegionService)
 	addressUsecase := NewAddressUsecase(mockAddressRepo, mockRegionService)
 
-	userID := uint(1)
+	userID := uint64(1)
 	req := &domain.CreateAddressRequest{
 		Name:       "Rumah",
 		Detail:     "Jl. Sudirman No. 123",
@@ -89,8 +89,8 @@ func TestAddressUsecase_GetAddressByID_Success(t *testing.T) {
 	mockRegionService := new(mocks.MockRegionService)
 	addressUsecase := NewAddressUsecase(mockAddressRepo, mockRegionService)
 
-	addressID := uint(1)
-	userID := uint(1)
+	addressID := uint64(1)
+	userID := uint64(1)
 	address := &domain.Address{
 		ID:     addressID,
 		UserID: userID,
@@ -119,8 +119,8 @@ func TestAddressUsecase_GetAddressByID_AccessDenied(t *testing.T) {
 	mockRegionService := new(mocks.MockRegionService)
 	addressUsecase := NewAddressUsecase(mockAddressRepo, mockRegionService)
 
-	addressID := uint(1)
-	userID := uint(2)
+	addressID := uint64(1)
+	userID := uint64(2)
 
 	// Mock expectations
 	mockAddressRepo.On("CheckOwnership", addressID, userID).Return(false)
@@ -142,8 +142,8 @@ func TestAddressUsecase_DeleteAddress_Success(t *testing.T) {
 	mockRegionService := new(mocks.MockRegionService)
 	addressUsecase := NewAddressUsecase(mockAddressRepo, mockRegionService)
 
-	addressID := uint(1)
-	userID := uint(1)
+	addressID := uint64(1)
+	userID := uint64(1)
 
 	// Mock expectations
 	mockAddressRepo.On("CheckOwnership", addressID, userID).Return(true)
